// javascript/packages/nextjs/src/internal/proto/action.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.0
//   protoc               v5.29.0
// source: action.proto

/* eslint-disable */

export const protobufPackage = "waymark.action";

/** Runtime required to execute an action. */
export enum ActionRuntime {
  ACTION_RUNTIME_UNSPECIFIED = 0,
  ACTION_RUNTIME_PYTHON = 1,
  ACTION_RUNTIME_JAVASCRIPT = 2,
  UNRECOGNIZED = -1,
}

export function actionRuntimeFromJSON(object: any): ActionRuntime {
  switch (object) {
    case 0:
    case "ACTION_RUNTIME_UNSPECIFIED":
      return ActionRuntime.ACTION_RUNTIME_UNSPECIFIED;
    case 1:
    case "ACTION_RUNTIME_PYTHON":
      return ActionRuntime.ACTION_RUNTIME_PYTHON;
    case 2:
    case "ACTION_RUNTIME_JAVASCRIPT":
      return ActionRuntime.ACTION_RUNTIME_JAVASCRIPT;
    case -1:
    case "UNRECOGNIZED":
    default:
      return ActionRuntime.UNRECOGNIZED;
  }
}

export function actionRuntimeToJSON(object: ActionRuntime): string {
  switch (object) {
    case ActionRuntime.ACTION_RUNTIME_UNSPECIFIED:
      return "ACTION_RUNTIME_UNSPECIFIED";
    case ActionRuntime.ACTION_RUNTIME_PYTHON:
      return "ACTION_RUNTIME_PYTHON";
    case ActionRuntime.ACTION_RUNTIME_JAVASCRIPT:
      return "ACTION_RUNTIME_JAVASCRIPT";
    case ActionRuntime.UNRECOGNIZED:
    default:
      return "UNRECOGNIZED";
  }
}